Add files via upload
This is the full tutorial in English explaining the theory, steps, references to coding files (uploaded in the "Contador" directory), and helpful tips for users. This tutorial aims to offer tools and support to other biologists who are not familiar with computer science so that they can adapt our pipeline to their own needs and research questions.